login user

+++ resolvers/booking.js
@@ -2,22 +2,28 @@ const Booking = require("../models/booking");
 
 const { transformBooking } = require("./merge");
 module.exports = {
-  createBooking: async ({ bookingInput }) => {
+  createBooking: async ({ bookingInput }, req) => {
     try {
+      if (!req.isAuth) {
+        throw Error("User not Authenticated");
+      }
       const res = await Booking.create({
         Event: bookingInput.eventId,
-        user: "65151ed8e3d7bed2404ee043",
+        user: req.userId,
       });
       return transformBooking(res, {
         Event: bookingInput.eventId,
-        user: "65151ed8e3d7bed2404ee043",
+        user: req.userId,
       });
     } catch (error) {
       throw error;
     }
   },
-  cancelBooking: async ({ bookingInput }) => {
+  cancelBooking: async ({ bookingInput }, req) => {
     try {
+      if (!req.isAuth) {
+        throw Error("User not Authenticated");
+      }
       const res = await Booking.findById(bookingInput.eventId);
       await Booking.findOneAndDelete(bookingInput.eventId);
       return transformBooking(res, {
@@ -28,7 +34,10 @@ module.exports = {
       throw error;
     }
   },
-  bookings: () => {
+  bookings: (args, req) => {
+    if (!req.isAuth) {
+      throw Error("User not Authenticated");
+    }
     return Booking.find()
       .then((data) => {
         return data?.map((booking) =>
